Fix losing focus problem in inline-task-editor
It's caused by a backend id that overrides the id of the div, also the key, forcing it to rerender. In this process, the ref gets invalidated.
Moral: Use ref only if you absolutely need to. We must use ref here because focus management can only be done by ref.