Put repeating task behind a flag
This commit continues the deployment preparation starting from #277
By default, it will not be enabled and won't be accidently enabled because you have to set local storage in console manually.
To check I doesn't remove the feature completely, run `localStorage.set('REPEATING_TASK_ENABLED', 'true')` and refresh, the feature will reappear. Unset it, it will disappear. It's expected that this flag will be removed once we deployed a stable version to production.