Add script to add `type: 'ONE_TIME'` to all existing task (#280)
To deploy a new version to production, we first need to normalize the database to  use the latest schema with the additional type field. This script does the normalization.

Test Plan:
To ease your testing, changing 500 to 2 in the code.
1. Create 5-6 random tasks on the staging site.
2. In firebase console, remove the type field.
3. Download firebase admin sdk of the STAGING site and put it in admin-scripts/firebaseadmin.json
4. `python3 add-type-one-time.py`
5. Observe the firebase console. The field should be added back.

I have run this on my machine, but it's highly recommended that a code reviewer also run this to ensure we don't mess up the production db.