[ci-cd] Setup GitHub actions to check, build, and deploy PR changes
Summary:
Why we need this: We want to see pr changes get deployed so that we can directly play with it without installing anything.
Why is this complicated: We want deployment to happen in isolation with other PR so that they don't interfere
Strategy: push to a subdirectory in GitHub pages so that it's isolated between branches and it still has the same domain so Firebase auth can still work.

Test Plan:
See CI build