Fix bottom bear not shown bug in mobile safari
This is a hack. I don't like it.
If Safari is a more decent browser than I don't have to do some thing like this.
The 100vh is causing trouble because mobile safari is deceptive: https://medium.com/@susiekim9/how-to-compensate-for-the-ios-viewport-unit-bug-46e78d54af0d
The easiet solution is just to make root fixed and span the entire visible page.
Mobile safari should just die.