Start to unify usage of current time for deterministic testing (#328)
### Summary

Last time I tried to add more snapshot tests to increase the test coverage of samwise. The journey stops at `FutureView` because it uses `new Date()` all over the place. Using `new Date()` prevents effective snapshot testing because a snapshot taken on date A will fail on date B.

Unifying the usage of current time is a necessary step to solve the issue. If we can shield all usage of current time behind a framework, then we can easily dig some hole in the framework for overriding the system time with our hardcoded time for deterministic testing, as shown in this diff.

Unifying the usage also has another advantage: when we plan to support more specific time like `2019-10-30 17:00` in the future, we don't have to look all over the codebase to hunt down the usage of `new Date()` because it has already be resolved when we are pushing more deterministic time.

Surprisingly, it has the third advantage: now we can get rid of `react-live-clock` which does not have a typescript type definition. One less eslint warning!

This diff does not plan to solve all problems. More practices are left to our amazing members!

### Test Plan

- Added snapshot tests for the entire FutureView
- Manual Test: open the page, change your system time, now the live clock and future view layout should change!