Better search box (#332)
* Better search box

### Summary

The `react-search-box` dependency is the latest dependency that does not have a typescript definition and it's the one that blocks us from having zero eslint warnings. The component is buggy and has poor customization ability, and in the past, we have to do ugly workarounds to overcome that.

Therefore, let's just write a new one, with all the unnecessary stuff removed! This rewrite uses React hooks, allows better style customization, and actually rerender itself when needed instead of using the `key` hack.

Now our eslint warnings are finally down to zero, and in this process I am also able to get rid of some `!important` in css.

### Test Plan

- [x] Snapshot tests have been updated to reflect the change.
- [x] Manual test: go to settings, search `CS 2112`. Compare that with current staging site for master.