Release new version to production (#347)
* Fix misleading/default tooltip values

* Add space before closing tag

* Update snapshots

* Add strikethrough for completed tasks in focus view

* Fix linter error

* Format line

* [editor][misc] Hide lastpass form fill overlay (#322)

This overlay is very annoying!

* [ci] Fix cleanup workflow (#324)

### Summary

The cleanup workflow doesn't actually cleanup. Try to fix it according to: https://github.community/t5/GitHub-Actions/Running-a-job-on-pull-request-merge-in-specific-branch/m-p/34192#M1829

### Test Plan

I will close the PR to see what happens.

* [misc] Fix weird indentation in firestore.rules (#325)

### Summary

The changed line appears to have a mix of tabs and space before 🤔. Not sure what happened. Now unifying it to 2-space to be consistent with the rest of the codebase.

### Test Plan

eyes

* added styles for special end dates on repeating tasks, the corresponding buttons, and added that css styling to the <p> containing last day of ____ buttons under repeating tasks

* Start to unify usage of current time for deterministic testing (#328)

### Summary

Last time I tried to add more snapshot tests to increase the test coverage of samwise. The journey stops at `FutureView` because it uses `new Date()` all over the place. Using `new Date()` prevents effective snapshot testing because a snapshot taken on date A will fail on date B.

Unifying the usage of current time is a necessary step to solve the issue. If we can shield all usage of current time behind a framework, then we can easily dig some hole in the framework for overriding the system time with our hardcoded time for deterministic testing, as shown in this diff.

Unifying the usage also has another advantage: when we plan to support more specific time like `2019-10-30 17:00` in the future, we don't have to look all over the codebase to hunt down the usage of `new Date()` because it has already be resolved when we are pushing more deterministic time.

Surprisingly, it has the third advantage: now we can get rid of `react-live-clock` which does not have a typescript type definition. One less eslint warning!

This diff does not plan to solve all problems. More practices are left to our amazing members!

### Test Plan

- Added snapshot tests for the entire FutureView
- Manual Test: open the page, change your system time, now the live clock and future view layout should change!

* Clean Up Readme

* Update Readme Roles

* Include all roles

* Wording fix

* Remove undo

* Prompt confirmation when deleting non-forked task

* Remove noUndo from arguments

* Fix linter error

* Set prompt message based on isFork

* Fix linter

* Better search box (#332)

* Better search box

### Summary

The `react-search-box` dependency is the latest dependency that does not have a typescript definition and it's the one that blocks us from having zero eslint warnings. The component is buggy and has poor customization ability, and in the past, we have to do ugly workarounds to overcome that.

Therefore, let's just write a new one, with all the unnecessary stuff removed! This rewrite uses React hooks, allows better style customization, and actually rerender itself when needed instead of using the `key` hack.

Now our eslint warnings are finally down to zero, and in this process I am also able to get rid of some `!important` in css.

### Test Plan

- [x] Snapshot tests have been updated to reflect the change.
- [x] Manual test: go to settings, search `CS 2112`. Compare that with current staging site for master.

* Remove toast dependency

* Only Search on Long Inputs

* Add No-Space Entry to Search

* Move all cancel buttons to the right of modal

* Add margin and padding to button

* Add overflow-y style and maxHeight in FloatingTaskEditor

* Fix linter error (singlequotes)

* Try to unbreak CI staging deploy (#336)

### Summary

Using this well-maintained github action instead of writing git magic on our own:
https://github.com/peaceiris/actions-gh-pages

If you see the source of that github action, it does some more sophisticated magic than mine original git scripts, which is probably the reason why using this action fixed the problem.

### Test Plan

If it works, `gh-pages` branch will show deployment status again.

It works: https://github.com/cornell-dti/samwise/commit/80716749153d3cc74ad26f87bbb96d99019baa45
See the green check. Previous commits do not have that.

* Move save button overflow styles to css

* Move all cancel buttons to the left

* set new date to midnight when you make a repeating task

* potentially fixed repeating task start/end but broke the app

* removed log

* change the enddate back to strictly > and instead add the end date as 11:59:999 PM the previous day.

* Add tabbing functionality to search bar

* Add arrow key handler for possible solution to arrow key support

* Change Repeating Task Icon to Designers' Icon

* fix linter errors

* Fix linter errors and revert dependency commit

* moved datetime math to util/hooks and added tests for the math

* Fix merge conflict

* Fix linebreak issue on Windows

* Hack CSS to make icon white

* Update snapshots to match fixes

* Added space