Actually fix the save button get hidden issue.
Here is the root cause of the problem: save button is hidden initially when there is no change. When there is a change, save button will appear, changing the height as a result. However, this height change is never detected by `FloatingTaskEditor`.

This commit tries to make the `FloatingTaskEditor` be aware of the height change. Luckily, we don't to write additional DOM manipulation code to achieve that. The height changes when editor content changes, so we can simply make `FloatingTaskEditor` be aware of content change, and `FloatingTaskEditor` can choose to call `updateFloatingEditorPosition` to adjust the position.

I decided to do the change detection in the reducer to avoid polluting the rest of the code with this implementation detail.