Improve performance for useMappedWindowSize hook.
### Summary

Problem: right now resizing window is very expensive.

Whenever the size changes, the listener will be triggered. We have a bunch of places that call the hook; however, most of the places are only interested in whether the window size means a desktop or mobile view, in other words, a mapped value.

A mapped value changes less often than window size. Right now, we are unconditionally rerendering the view even if the mapped value does not change. This PR aims to change that.

This PR duplicates some code so that we will only call the `setMappedValue` function only if the mapped value actually change. In this way, we can avoid unnecessary rerenders.

### Test Plan

Try to resize. Nothing breaks. Things should be much faster.