Fix checking unsaved subtask. Close #354 (#380)
### Summary

This bug is introduced a while ago while we added autosave back in #309.
The fix is simple: save the edit to the local reducer first. The call of `editSubTask` will just fail silently and harmlessly.

### Test Plan

Create a task, type one character in add subtask box and click checkbox very quickly. It's indeed saved.