Unbreak course scraper for SP20 (#393)
### Summary

This diff simply unbreaks it. Now it can generate a JSON we want. I forgot what's the process of putting the JSON to the right place and ensure there is no compatibility problems. Let's leave that to today's work session.

In the long term, we should think about how we can make this more maintainable. I'm happy to rewrite the script in TS and move my `cornell-api-libs` to DTI org.

### Test Plan

```bash
cd course-info
./run
# no errors, only warnings are printed.
```