Add an integration test for course info scraper (#404)
* Add an integration test for course info scraper

### Summary

I just thought about it today and found that we can easily write an integration test to test whether the latest course-info json is consistent with the one on the master. We need this check to ensure that our information is up-to-date.

This check proves to be necesarry, since I do found some inconsistencies, probably because some courses and exams are added. The course json has been updated to reflect the latest state.

### Test Plan

See the CI job yml. Note that I used a tool called `jq` to normalize the json. It can remove potential inconsistency caused by json entries order and ensure that in case there is an inconsistency, the diff outputed is more readable.

* Run Yarn install before running `./course-info/run`

The scraper has some dependencies. We need to install them first!

* Force CI to new New York timezone

* Fix setting env-var syntax

* Ensure the script gets the updated TZ envvar