Automatically create PR to update JSON when E2E course-info failed (#411)
### Summary

This pull request tries to let @dti-github-bot automatically create a PR when the E2E test on course info json failed. I hope this can reduce our burden to update stuff.
Ideally, we can show the diff inline in the created PR comment, but it might us to switch the entire workflow to github actions. It's not too bad to look at the latest master E2E test failure to see what's changed, so I decide not to implement it, as least in this diff.

### Test Plan

See CI running result. The PR is created in https://github.com/cornell-dti/samwise/pull/412.
The consistency test is designed to fail, so that the PR creation job can be triggered. The create PR job also fails because the PR is already created. This error cannot be try-catched, but I don't think it's too big because it can remind people to review PRs. :)