Automatically remove old non-repeating tasks (#418)
* Automatically remove old non-repeating tasks

### Summary

This diff adds a function that removes all old non-repeating tasks.
The function currently uses 40 as the upper limit to try to be conservative, because if we allow user to see tasks a month ago, it can still be supported.

While testing the function, I verified that Michael's report that the deploy script is indeed broken, because I was using an old globally installed firebase CLI before. It's fixed in this diff. To prevent similar problems in the future, I integrated firebase functions deployment into continuous deployment script.

### Test Plan

Deployed to staging to see it's effect. Looks like it works.

I also write a unit test for the `partition` helper function, and the test is run by `react-scripts`. Unfortunately `react-scripts` automatically overrides some json fields in tsconfig.json, so I have to override it back in tsc CLI in package.json. The alternative would be setting up jest manually, but it requires an additional jest and an additional babel config, and I hate writing those configs.

* Create dummy credential in CI test environment

* Remove python script that does the same thing

The python script can be run only manually, while the function is automatically run every day.

* Skip initialize firebase-admin in test mode