Reduce course json update frequency (#427)
Even if it's checking twice a day, we don't review code that often and sometimes we will batch json updates. Let's reduce the frequency to be only run on Mon/Wed/Fri to reduce the noise. I choose workdays because part json update record shows that cornell does not update these information on weekends.