Make build and test in frontend work across different OS (#428)
* Make build and test in frontend work across different OS

### Summary

Instead of hardcoding envvar in the script, put them in a env file and load it by cross-platform node.js code.

### Test Plan

Testing it locally. It works. See CI log to verify that it also works in CI.
Can a reviewer on Windows help me test whether it works on Windows?

### Notes

Inspired by https://create-react-app.dev/docs/deployment#customizing-environment-variables-for-arbitrary-build-environments

* Remove snapshot testing on titlebar to remove need to set `TZ=UTC`