Task Type Refactor
### Summary

This diff continues the effort made in #423. In this diff, we are finally able to group the content of the task together, and factor all one-time task and repeating task specific stuff to `TaskMetadata`. In the future, we can add more stuff to task metadata without greatly impacting the rest of our codebase.

This diff is intended to be the first step of cleaning up our frontend representation of task. It does not plan to cleanup the rest of the code. The only goal of this diff is to refactor the type, make the app compile, and ensure nothing breaks.

The only potentially controversial change is to move `date` into `TaskMetadata`. While it might seem odd because a task has a date is so central to the app, I believe it is necessary because the behavior is fundamentally different for one-time and repeating task. Giving repeating task a dummy date will make buggy code that makes bad assumptions of date escape through the type checker.

### Test Plan

- [x] Changed existing tests due to type refactoring.
- [x] Manually played with the app. Everything else still seems to work.