Disallow editing/deleting canvas task again
### Summary

During the refactoring of #427, we moved all non-common fields of tasks into the metadata object. The transformation is done in `src/firebase/listeners.ts`. The transformer forgets to put the `icalUID` field into `metadata` object, which makes a lot of canvas task check break.

The reason we didn't catch the bug or didn't notice it is that the field is not even declared in its firebase types! Now we declare the field and fix the transformer.

### Test Plan

Canvas tasks cannot be deleted anymore. See the screenshot.