Seamless drag-and-drop experience (#493)
* Seamless drag-and-drop experience

The drag and drop user experience is far from being perfect.
To make order persistent, we have to save the updated order in the firestore. Although firestore seems to have ['latency compensation'](https://firebase.google.com/docs/firestore/query-data/listen#events-local-changes) that can avoid a roundtrip before updated data is pushed to remote, the listener update is still not **synchronous**, which causes a big trouble in dnd that relies an a synchronous state change.

In the past, we introduced `localTasks` as a cache in focus view to solve the problem. We are able to mostly avoid the dropped task bouncing back to the original position, but there are still some occasional flashes. In future view, such caching mechanism has not been implemented, so the resulting experience is even worse.

Recently, I encountered a similar problem in my side projects, and I found a more principled and easier way to solve it. Surprisingly, this easier way also makes the UX better. An crucial observation is that

> Redux store update is synchronous.

Therefore, instead of `setState` locally in a component, why don't we do similar stuff to the redux store? Then we can avoid creating two versions of truth. It's also easier to implement, since the functions we call to update on firebase already works on change sets, and our redux store updates also work on change set.

After this change, we are able to completely remove those ugly local caches and make the app more responsive!

Try to drag and drop within focus view, from focus view completed to uncompleted, within a day in future view and between days.
No more flashes!

* Convert != null check to !== undefined check

To be more precise