Eliminate all absolute imports
Absolute imports can be very confusing because the starting folder sometimes shares the same with the library we are using. e.g. We have a top-level folder `firebase` in src and also `firebase` as a dependency. When we import `firebase/xyz`, it's not immediately clear whether we are importing `firebase` from a local util or from `firebase` in node_modules.

This diff removes the `baseUrl` in tsconfig.json, and use that as a guide to remove all remaining absolute imports.