Fix group tasks icon in personal view
## Summary

Originally, we have the default icon next to the task title be the repeating icon, so now when we add a new type `GROUP`, it will default to the repeating icon. Let's make the Placeholder (empty) the default and if it is `ONE_TIME` and not a canvas task, we can show the `DragIcon`. All other cases, leave the icon blank.

## Test Plan