Move blanket button styling of settings page to SettingButton css (#591)
We shouldn't do this kind of blanket button styling at all, since it affects too many buttons at once. However, we are already stuck with this for a long time and it's already affecting too many buttons. :(

This diff tries to mitigate the problem by moving two same instances of this blanket button styling into 1SettingButton.module.css`, so that this kind of bad styling practice only appears once, so it will be easier to us to fix them in the future. It also helps to remove another `no-descending-specificity` css warning.