Loosen firestore DataConverter runtime type checking with timestamp (#718)
Currently, the functions code is failing, because the runtime type checking asserts the field is an instance of firestore.Timestamp. However, this class is frontend only and in functions it's using another class. Unfortunately, there is no easy way to make frontend and backend share these runtime type checking code, so I decide to loosen it to `any`, so that at least functions don't break.