Remove dependencence on store store in firestore common actions (#722)
These actions should be able to be called on functions, which doesn't have a redux store.

Therefore, I make the common actions to take full task as parameters instead of task id. Then in the frontend actions, I pass in the task fetched from redux store.

The methods refactored in this diff is important to implement #333.